Refactored how systems are installed
The previous changes to where galaxies are installed forced changes to how systems are installed. The `ANSIBLE_ROLES_PATH` environment variable is used to tell Ansible where to find roles, which may be in the `matter` directory found in any galaxy. To find the system to install, the galaxies are iterated until the first name match is found. If there are name conflicts, then the first one wins.