Support for installing systems with fully-qualified names
`bang user/repo/system` will look specifically for a system in `/usr/local/Galaxies/[user]/[repo]/systems/` that is named `[system].yml`. If that system does not exist, then it will fail. If a fully-qualified name is not used, then the standard practice of searching all galaxies for the first matching system will be used.