ansible-playbook output is redirected to the Ruby process
Using `Kernel.system` allows for showing real-time output instead of waiting for the entire process to finish. Both stdout and stderr are redirected. It's `:chdir` option makes the call cleaner since we don't have to `cd BANG_LIB &&` before `ansible-playbook`. And now that stdout is redirected, use of Ansible's pause module with the prompt parameter enables prompting the user for input before proceeding with play execution.